CRONAPP-3229 - Ajuste Signup (#308)
https://cronapp.atlassian.net/browse/CRONAPP-3329

Problema:
Não existia bloco para reset password

Solução:
Desenvolvido bloco para o reset password
