+++ project/W/cronapp-rad-project/src/main/webapp-autenticacao/webapp/views/login.view.html.ftl
@@ -151,7 +151,6 @@
     </div>
 </div>
 
-
 <div class="modal fade" id="forgotPasswordModal">
     <div class="modal-dialog">
         <div class="modal-content">
@@ -173,20 +172,16 @@
                             <label for="forgotPasswordEmail">
                                 {{'ForgotPasswordEmail' | translate}}
                             </label>
-                            <input type="email" class="form-control k-textbox" id="forgotPasswordEmail"
-                                   aria-label="{{'ForgotPasswordEmail' | translate}}" ng-required="true"
-                                   ng-model="forgotPasswordEmail"/>
+                            <input type="email" class="form-control k-textbox" id="forgotPasswordEmail" aria-label="{{'ForgotPasswordEmail' | translate}}" ng-required="true" ng-model="forgotPasswordEmail">
                         </div>
                     </div>
                 </form>
             </div>
             <div class="modal-footer">
-                <button class="btn btn-primary btn-fab k-button" type="submit" ng-click="forgotPassword()"
-                        ng-disabled="form.$invalid || vm.dataLoading" aria-label="{{'Save' | translate}}">
+                <button class="btn btn-primary btn-fab k-button" type="submit" ng-click="cronapi.client('js.blockly.UserControl.resetPassword').names('7afdf898').run(forgotPasswordEmail)" ng-disabled="form.$invalid || vm.dataLoading" aria-label="{{'Save' | translate}}">
                     <span class="k-icon k-i-check"></span>
                 </button>
-                <button class="btn btn-default btn-fab k-button" type="button" data-dismiss="modal"
-                        aria-label="{{'Home.view.Close' | translate}}">
+                <button class="btn btn-default btn-fab k-button" type="button" data-dismiss="modal" aria-label="{{'Home.view.Close' | translate}}">
                     <span class="k-icon k-i-close"></span>
                 </button>
             </div>
